+++ src/platform/wasm/constraints.rs
@@ -198,8 +198,8 @@ impl From<DisplayAudioTrackConstraints> for MediaTrackConstraints {
     fn from(_: DisplayAudioTrackConstraints) -> Self {
         let this = Self::new();
 
-        // TODO: Consider allowing to toggle audio processing settings,
-        //       but that would require medea-flutter-webrtc changes.
+        // TODO: Consider allowing to toggle audio processing settings.
+        //       That would require medea-flutter-webrtc changes.
         // Disable audio processing for system audio on web, this aligns
         // with current desktop implementation in medea-flutter-webrtc
         // which has no APM for system audio at all.
@@ -214,6 +214,7 @@ impl From<DisplayAudioTrackConstraints> for MediaTrackConstraints {
         ));
 
         // TODO: Not implemented in `web_sys` so have to use reflect.
+        // Excludes current tab audio. Supported on Chrome 141+.
         drop(Reflect::set_with_receiver(
             &this,
             &"restrictOwnAudio".into(),