+++ CHANGELOG.md
@@ -6,9 +6,12 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
-## master
+## [0.13.0] · 2026-01-19
+[0.13.0]: https://github.com/instrumentisto/medea-jason/tree/medea-jason-0.13.0
 
-[Diff](https://github.com/instrumentisto/medea-jason/compare/medea-jason-0.12.0...master)
+[Diff](https://github.com/instrumentisto/medea-jason/compare/medea-jason-0.12.0...medea-jason-0.13.0)
+
+See also [`medea_jason` pub package 0.13.0 changes](https://github.com/instrumentisto/medea-jason/tree/medea-jason-0.13.0/flutter/CHANGELOG.md).
 
 ### Added